Refactor battle logging into service
